fp_hmm

This feature consists of two vectors, \(d, s\). Vector \(d\) corresponds to the sums across the sequence for each of the 20 amino acid columns. Vector \(s\) corresponds to a flattened matrix \(S[i, j] = \sum_{k = 1}^{L} H[k, j] \times \delta[k, i]\) in which \(\delta[k, i] = 1\) when \(A_i = H[k, j]\). \(A\) refers to a list of all possible amino acids, \(i, j\) span from \(1:20\).

Usage

fp_hmm(hmm)

Arguments

hmm

The name of a profile hidden markov model file.

Value

A vector of length 20.

A vector of length 400.

References

Zahiri, J., Yaghoubi, O., Mohammad-Noori, M., Ebrahimpour, R., & Masoudi-Nejad, A. (2013). PPIevo: Protein–protein interaction prediction from PSSM based evolutionary information. Genomics, 102(4), 237–242.

Examples

h<- fp_hmm(system.file("extdata", "1DLHA2-7", package="protHMM"))